Simplify plugin using ImageJ2 automatic detection
