+++ render-ws-java-client/pom.xml
@@ -164,6 +164,16 @@
         </dependency>
 
         <dependency>
+            <groupId>net.imagej</groupId>
+            <artifactId>imagej</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>net.imagej</groupId>
+            <artifactId>imagej-legacy</artifactId>
+        </dependency>
+
+        <dependency>
             <groupId>sc.fiji</groupId>
             <artifactId>z_spacing</artifactId>
             <exclusions>
